pass down compound service from top level component, cleanup
