+++ app/src/components/market/sections/market_buy/market_buy_container.tsx
@@ -1,11 +1,13 @@
 import React from 'react'
 
+import { CompoundService } from '../../../../services'
 import { MarketDetailsTab, MarketMakerData } from '../../../../util/types'
 
 import { MarketBuy } from './market_buy'
 import { ScalarMarketBuy } from './scalar_market_buy'
 
 interface Props {
+  compoundService: CompoundService
   isScalar: boolean
   marketMakerData: MarketMakerData
   switchMarketTab: (arg0: MarketDetailsTab) => void
