+++ app/src/components/market/sections/market_sell/market_sell_container.tsx
@@ -1,11 +1,13 @@
 import React from 'react'
 
+import { CompoundService } from '../../../../services'
 import { MarketDetailsTab, MarketMakerData } from '../../../../util/types'
 
 import { MarketSell } from './market_sell'
 import { ScalarMarketSell } from './scalar_market_sell'
 
 interface Props {
+  compoundService: CompoundService
   isScalar: boolean
   marketMakerData: MarketMakerData
   switchMarketTab: (arg0: MarketDetailsTab) => void
